+++ settei/base.py
@@ -4,7 +4,9 @@
 .. versionadded:: 0.2.0
 
 """
+import collections
 import collections.abc
+import enum
 import functools
 import pathlib
 import re
@@ -92,7 +94,7 @@ class config_property:
     """
 
     @typechecked
-    def __init__(self, key: str, cls, docstring: str=None, **kwargs) -> None:
+    def __init__(self, key: str, cls, docstring: str = None, **kwargs) -> None:
         self.key = key
         self.cls = cls
         self.__doc__ = docstring
@@ -119,11 +121,12 @@ class config_property:
             self.default_func = None
             self.default_warning = False
 
-    def __get__(self, obj, cls: typing.Optional[type]=None):
+    def __get__(self, obj, cls: typing.Optional[type] = None):
         if obj is None:
             return self
         default, value = self.get_raw_value(obj)
         if not default:
+            value = self.convert_native_type(value)
             self.typecheck(value)
         return value
 
@@ -146,6 +149,45 @@ class config_property:
                     return True, default
                 raise ConfigKeyError(key)
         return False, value
+
+    def convert_native_type(self, value) -> typing.Any:
+        cls = get_union_types(self.cls) or self.cls
+        if isinstance(cls, type) and issubclass(cls, enum.Enum):
+            try:
+                return cls(value)
+            except ValueError:
+                raise ConfigTypeError(
+                    'Invalid value {0} in {1!r}. Candidates are: {2}'.format(
+                        value, cls, ', '.join(cls.__members__)
+                    )
+                )
+        elif isinstance(cls, collections.Iterable):
+            enums = filter(lambda i: issubclass(i, enum.Enum), cls)
+            non_enums = filter(lambda i: not issubclass(i, enum.Enum), cls)
+            candidates = []
+            for e in enums:
+                try:
+                    candidates.append(e(value))
+                except ValueError:
+                    pass
+            if not candidates:
+                if non_enums:
+                    return value
+                else:
+                    raise ConfigTypeError(
+                        'No matching value {0} for types: {1}'.format(
+                            value, ', '.join([repr(r) for r in enums])
+                        )
+                    )
+            elif len(candidates) == 1:
+                return candidates[0]
+            else:
+                raise ConfigTypeError(
+                    'Ambiguous enum type for value {0}: {1}'.format(
+                        value, ', '.join([repr(r) for r in candidates])
+                    )
+                )
+        return value
 
     def typecheck(self, value) -> None:
         union_types = get_union_types(self.cls)
@@ -280,12 +322,12 @@ class config_object_property(config_property):
     )
 
     @typechecked
-    def __init__(self, key: str, cls, docstring: str=None, recurse: bool=False,
-                 **kwargs) -> None:
+    def __init__(self, key: str, cls, docstring: str = None,
+                 recurse: bool = False, **kwargs) -> None:
         super().__init__(key=key, cls=cls, docstring=docstring, **kwargs)
         self.recurse = recurse
 
-    def __get__(self, obj, cls: typing.Optional[type]=None):
+    def __get__(self, obj, cls: typing.Optional[type] = None):
         if obj is None:
             return self
         default, expression = self.get_raw_value(obj)
@@ -431,7 +473,7 @@ class Configuration(collections.abc.Mapping):
             return cls.from_file(f)
 
     @typechecked
-    def __init__(self, config: typing.Mapping[str, object]={}, **kwargs):
+    def __init__(self, config: typing.Mapping[str, object] = {}, **kwargs):
         self.config = dict(config, **kwargs)
 
     def __len__(self) -> int: